refactor!: ship ESM-only builds
