--- .github/publish-ci/cra5/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "rs-esm-cra",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@react-spring/web": "*",
-    "react": "^19.2.6",
-    "react-dom": "^19.2.6",
-    "react-scripts": "5.0.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject",
-    "test": "echo Done"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "msw": {
-    "workerDirectory": "public"
-  },
-  "devDependencies": {
-    "@types/node": "24.12.4",
-    "@types/react": "^19.2.15",
-    "@types/react-dom": "^19.2.3",
-    "typescript": "4.9.5"
-  }
-}
