--- .github/publish-ci/cra5/src/App.css
@@ -1,6 +1,0 @@
-.spring {
-  width: 80px;
-  height: 80px;
-  background: #00ff00;
-  border-radius: 16px;
-}
