+++ .github/publish-ci/node-esm/package.json
@@ -5,14 +5,11 @@
   "type": "module",
   "scripts": {
     "build": "echo Done",
-    "test": "node test.cjs"
+    "test": "node test.js"
   },
   "dependencies": {
     "@react-spring/web": "*",
     "react": "^19.2.6",
     "react-dom": "19.2.6"
-  },
-  "devDependencies": {
-    "resolve-esm": "2.0.3"
   }
 }
