+++ packages/parallax/package.json
@@ -1,20 +1,14 @@
 {
   "name": "@react-spring/parallax",
   "version": "10.1.1",
-  "module": "./dist/react-spring_parallax.legacy-esm.js",
-  "main": "./dist/cjs/index.js",
+  "type": "module",
+  "module": "./dist/react-spring_parallax.modern.mjs",
   "types": "./dist/react-spring_parallax.modern.d.mts",
   "exports": {
     "./package.json": "./package.json",
     ".": {
-      "import": {
-        "types": "./dist/react-spring_parallax.modern.d.mts",
-        "default": "./dist/react-spring_parallax.modern.mjs"
-      },
-      "require": {
-        "types": "./dist/cjs/react-spring_parallax.development.d.ts",
-        "default": "./dist/cjs/index.js"
-      }
+      "types": "./dist/react-spring_parallax.modern.d.mts",
+      "default": "./dist/react-spring_parallax.modern.mjs"
     }
   },
   "files": [
