+++ packages/rafz/package.json
@@ -2,20 +2,14 @@
   "name": "@react-spring/rafz",
   "version": "10.1.1",
   "description": "react-spring's fork of rafz one frameloop to rule them all",
-  "module": "./dist/react-spring_rafz.legacy-esm.js",
-  "main": "./dist/cjs/index.js",
+  "type": "module",
+  "module": "./dist/react-spring_rafz.modern.mjs",
   "types": "./dist/react-spring_rafz.modern.d.mts",
   "exports": {
     "./package.json": "./package.json",
     ".": {
-      "import": {
-        "types": "./dist/react-spring_rafz.modern.d.mts",
-        "default": "./dist/react-spring_rafz.modern.mjs"
-      },
-      "require": {
-        "types": "./dist/cjs/react-spring_rafz.development.d.ts",
-        "default": "./dist/cjs/index.js"
-      }
+      "types": "./dist/react-spring_rafz.modern.d.mts",
+      "default": "./dist/react-spring_rafz.modern.mjs"
     }
   },
   "files": [
