+++ packages/shared/package.json
@@ -2,20 +2,14 @@
   "name": "@react-spring/shared",
   "version": "10.1.1",
   "description": "Globals and shared modules",
-  "module": "./dist/react-spring_shared.legacy-esm.js",
-  "main": "./dist/cjs/index.js",
+  "type": "module",
+  "module": "./dist/react-spring_shared.modern.mjs",
   "types": "./dist/react-spring_shared.modern.d.mts",
   "exports": {
     "./package.json": "./package.json",
     ".": {
-      "import": {
-        "types": "./dist/react-spring_shared.modern.d.mts",
-        "default": "./dist/react-spring_shared.modern.mjs"
-      },
-      "require": {
-        "types": "./dist/cjs/react-spring_shared.development.d.ts",
-        "default": "./dist/cjs/index.js"
-      }
+      "types": "./dist/react-spring_shared.modern.d.mts",
+      "default": "./dist/react-spring_shared.modern.mjs"
     }
   },
   "files": [
