+++ packages/types/package.json
@@ -2,20 +2,14 @@
   "name": "@react-spring/types",
   "version": "10.1.1",
   "description": "Internal package with TypeScript stuff",
-  "module": "./dist/react-spring_types.legacy-esm.js",
-  "main": "./dist/cjs/index.js",
+  "type": "module",
+  "module": "./dist/react-spring_types.modern.mjs",
   "types": "./dist/react-spring_types.modern.d.mts",
   "exports": {
     "./package.json": "./package.json",
     ".": {
-      "import": {
-        "types": "./dist/react-spring_types.modern.d.mts",
-        "default": "./dist/react-spring_types.modern.mjs"
-      },
-      "require": {
-        "types": "./dist/cjs/react-spring_types.development.d.ts",
-        "default": "./dist/cjs/index.js"
-      }
+      "types": "./dist/react-spring_types.modern.d.mts",
+      "default": "./dist/react-spring_types.modern.mjs"
     }
   },
   "files": [
