+++ targets/web/package.json
@@ -1,20 +1,14 @@
 {
   "name": "@react-spring/web",
   "version": "10.1.1",
-  "module": "./dist/react-spring_web.legacy-esm.js",
-  "main": "./dist/cjs/index.js",
+  "type": "module",
+  "module": "./dist/react-spring_web.modern.mjs",
   "types": "./dist/react-spring_web.modern.d.mts",
   "exports": {
     "./package.json": "./package.json",
     ".": {
-      "import": {
-        "types": "./dist/react-spring_web.modern.d.mts",
-        "default": "./dist/react-spring_web.modern.mjs"
-      },
-      "require": {
-        "types": "./dist/cjs/react-spring_web.development.d.ts",
-        "default": "./dist/cjs/index.js"
-      }
+      "types": "./dist/react-spring_web.modern.d.mts",
+      "default": "./dist/react-spring_web.modern.mjs"
     }
   },
   "files": [
